requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
distro==1.9.0
Django==5.2.5
djangorestframework==3.16.1
grpcio==1.74.0
grpcio-tools==1.74.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.10.0
lxml==6.0.0
mysqlclient==2.2.7
numpy==2.3.2
openai==1.100.1
packaging==25.0
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
PyMySQL==1.1.1
pyOpenSSL==25.1.0
pyproj==3.7.2
python-docx==1.2.0
python-dotenv==1.1.1
requests==2.32.4
scipy==1.16.1
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wheel==0.45.1
zope.interface==7.2
pymysql
gunicorn
django-cors-headers==4.4.0
drf-nested-routers==0.94.2
